archive/pyproject.toml
[project]
name = "roboblog"
version = "0.1.0"
description = "Automated blog generation from GitHub commits using LLM"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.31.0",
    "pyyaml>=6.0.0",
    "python-dotenv>=1.0.0",
    "dspy-ai>=2.4.0",
    "jinja2>=3.1.0",
    "pelican[markdown]>=4.11.0.post0",
]

[project.scripts]
fetch-commits = "roboblog.fetch_commits:main"
generate-post = "roboblog.generate_post:main"
sync-jekyll-config = "roboblog.sync_jekyll_config:main"
process-human-posts = "roboblog.process_human_posts:main"
run-blog-update = "roboblog.run_blog_update:main"

[project.optional-dependencies]
dev = [
    "ruff>=0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]

[tool.uv.workspace]
members = [
    "pelican",
]
